refactor(auth): work around cypress test issue
Cypress is importing `DEV_JURISDICTION` by referencing a `.ts` file directly. This works because webpack is transpiling anyway, but the `certs.ts` file it was importing now imports the `tmp` package indirectly. This fails since the polyfilled `fs` package webpack provides doesn't have a `constants` exports, so we get "`O_CREAT` is not a property of undefined" or some such error. This commit works around the issue by putting the constants in a `constants.ts` that the test can reference.
