+++ apps/admin/integration-testing/cypress/support/auth.ts
@@ -3,7 +3,7 @@ import { methodUrl } from '@votingworks/grout';
 // Importing all of @votingworks/auth causes Cypress tests to fail since Cypress doesn't seem to
 // interact well with PCSC Lite card reader code
 // eslint-disable-next-line vx/no-import-workspace-subfolders
-import { DEV_JURISDICTION } from '@votingworks/auth/src/certs';
+import { DEV_JURISDICTION } from '@votingworks/auth/src/constants';
 // eslint-disable-next-line vx/no-import-workspace-subfolders
 import {
   mockCard,
