+++ libs/auth/scripts/program_system_administrator_java_card.ts
@@ -3,7 +3,7 @@ import { assert } from '@votingworks/basics';
 import { generatePin, hyphenatePin } from '@votingworks/utils';
 
 import { ResponseApduError } from '../src/apdu';
-import { DEV_JURISDICTION } from '../src/certs';
+import { DEV_JURISDICTION } from '../src/constants';
 import { getRequiredEnvVar } from '../src/env_vars';
 import { JavaCard } from '../src/java_card';
 import { constructJavaCardConfig } from '../src/java_card_config';
