+++ libs/auth/src/index.ts
@@ -1,4 +1,4 @@
-export { DEV_JURISDICTION } from './certs';
+export { DEV_JURISDICTION } from './constants';
 export * from './dipped_smart_card_auth_api';
 export * from './dipped_smart_card_auth';
 export * from './inserted_smart_card_auth_api';
