+++ libs/auth/src/java_card.ts
@@ -27,7 +27,6 @@ import {
 } from './card';
 import { CardReader } from './card_reader';
 import {
-  CERT_EXPIRY_IN_DAYS,
   constructCardCertSubject,
   constructCardCertSubjectWithoutJurisdictionAndCardType,
   parseCardDetailsFromCert,
@@ -57,6 +56,7 @@ import {
   RESET_RETRY_COUNTER,
   VERIFY,
 } from './piv';
+import { CERT_EXPIRY_IN_DAYS } from './constants';
 
 /**
  * The OpenFIPS201 applet ID
